+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Typings for parsed SVF materials
+  - Note: breaking change of the `parseMaterials` function signature
+
 ## [6.0.2] - 2019-09-24
 
 ### Fixed